feature/#26 응답 불일치 수정
